Refer to Trac #442. Enhancing the old patch (r3486) with a few additional error checks.
We should not try to acquire the dreg locks when the registration cache has not
been initialized. 

If a process already owns the lock (a remote possibility), it
should try to lock it again. 

Finally, we have also introduced two
variables - "alltoall_dreg_disable_threshold" and "alltoall_dreg_disable".
If we want to disable registration cache for alltoall alone, irrespective of the
system size, we can set MV2_ALLTOALL_DREG_DISABLE=0. The MV2_ALLTOALL_DREG_DISABLE_THRESHOLD
variable denotes the largest system size upto which we can allow alltoall to use
the registration cache. The default value for this variable is 1024. 


git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3503 09bc9535-d30e-0410-b1f7-d46b20a4725c