Remove unhandled rdma cm events
Some of these cases may need to be handled but since they are basically
no ops right now and not all versions of librdmacm export each of these
events we can let the default case handle them.

Later if we find that we need to explicily handle a case we need to
either update our dependancy on librdmacm if we cannot function without
or guard the new portion of code based on the librdmacm version if it is
optional.




git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3715 09bc9535-d30e-0410-b1f7-d46b20a4725c