We need the CFLAGS -D_XOPEN_SOURCE=600 for totalview to work on RHEL5 systems.
If we have the CFLAGS -D_XOPEN_SOURCE=600 on RHEL4 systems, our library will not
compile

But, the configure file configure.in (meant for RHEL5 systems) doesn't have this
flag, while configure.in.rhel4 (meant for RHEL4 systems) has this.

This should be the other way around. So changing this.



git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3800 09bc9535-d30e-0410-b1f7-d46b20a4725c